Change indexes into tile arrays to be stored as bytes to save memory.
The index value needs only be big enough to handle all defined terrain types. This is a low number so we can save memory by defining it as a byte. This particularly saves memory for the CustomTerrain field in the Map class, which defines a cell layer for the map using tile indexes, so we can reduce the size of that layer 4x as a result.

+++ OpenRA.Mods.RA/AI/HackyAI.cs
@@ -158,7 +158,7 @@ namespace OpenRA.Mods.RA.AI
 		bool enabled;
 		int ticks;
 
-		HashSet<int> resourceTypeIndices;
+		HashSet<byte> resourceTypeIndices;
 
 		RushFuzzy rushFuzzy = new RushFuzzy();
 
@@ -208,7 +208,7 @@ namespace OpenRA.Mods.RA.AI
 
 			random = new MersenneTwister((int)p.PlayerActor.ActorID);
 
-			resourceTypeIndices = new HashSet<int>(
+			resourceTypeIndices = new HashSet<byte>(
 				Map.Rules.Actors["world"].Traits
 				.WithInterface<ResourceTypeInfo>()
 				.Select(t => world.TileSet.GetTerrainIndex(t.TerrainType)));
